+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the `dom_query` crate will be documented in this file.
 
+## [Unreleased]
+
+### Added
+- Introduced the `NodeRef::strip_elements(&[&str])` method, which removes matched elements while retaining their children in the document tree.
+- Introduced the `Selection::strip_elements(&[&str])` method, which does the same operation as `NodeRef::strip_elements(&[&str])` but for every node in the `Selection`.
+
 ## [0.16.0] - 2025-03-09
 
 ### Added